# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TapasHub Business Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/stats:
    get:
      operationId: getOrderStats
      tags: [orders]
      summary: Order statistics (counts by status, today vs yesterday)
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Order statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderStats"

  # ─── Inventory ───────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
        - name: role
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, orderNumber, companyId, companyName, customerId, customerName, status, totalAmount, itemCount, channel, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        companyId:
          type: integer
        companyName:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, returned, refunded]
        totalAmount:
          type: number
        itemCount:
          type: integer
        channel:
          type: string
          enum: [shopify, shopdeck, baapstore, website, whatsapp, direct]
        shippingAddress:
          type: ["string", "null"]
        trackingNumber:
          type: ["string", "null"]
        courierName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    OrderInput:
      type: object
      required: [companyId, customerName, totalAmount, channel]
      properties:
        companyId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        customerId:
          type: integer
        totalAmount:
          type: number
        channel:
          type: string
          enum: [shopify, shopdeck, baapstore, website, whatsapp, direct]
        shippingAddress:
          type: string
        notes:
          type: string
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, returned, refunded]
        trackingNumber:
          type: string
        courierName:
          type: string
        notes:
          type: string
    OrderList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    OrderStats:
      type: object
      required: [total, pending, processing, shipped, delivered, cancelled, todayRevenue, yesterdayRevenue]
      properties:
        total:
          type: integer
        pending:
          type: integer
        processing:
          type: integer
        shipped:
          type: integer
        delivered:
          type: integer
        cancelled:
          type: integer
        returned:
          type: integer
        todayRevenue:
          type: number
        yesterdayRevenue:
          type: number

    # ─── Inventory / Products ──────────────────────────────────────
...
    User:
      type: object
      required: [id, name, email, role, companyIds, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, founder, director, finance, operations, hr, marketing, customer_support, warehouse, vendor, ca, employee, investor]
        companyIds:
          type: array
          items:
            type: integer
        avatarUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, founder, director, finance, operations, hr, marketing, customer_support, warehouse, vendor, ca, employee, investor]
        companyIds:
          type: array
          items:
            type: integer